Switch to an 8-bit buffer for the console

+++ kernel/console.h
@@ -16,7 +16,7 @@ typedef struct {
 
   int width, height;
   int offset;
-  uint16_t *buffer;
+  uint8_t *buffer;
   point_t cur;
 
   semaphore_t write_sem;
